+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/indicator/BackgroundColorValidationIndicator.java
@@ -33,8 +33,8 @@ import static java.util.Objects.requireNonNull;
 public final class BackgroundColorValidationIndicator implements ValidationIndicator {
 
 	@Override
-	public void enable(JComponent component, ObservableState valid, ObservableState warned) {
-		new Indicator(requireNonNull(component), requireNonNull(valid), requireNonNull(warned));
+	public void enable(JComponent component, ObservableState invalid, ObservableState warned) {
+		new Indicator(requireNonNull(component), requireNonNull(invalid), requireNonNull(warned));
 	}
 
 	private static final class Indicator {
@@ -47,37 +47,37 @@ public final class BackgroundColorValidationIndicator implements ValidationIndic
 		private @Nullable Color invalidBackgroundColor;
 		private @Nullable Color warnedBackgroundColor;
 
-		private final ObservableState valid;
+		private final ObservableState invalid;
 		private final ObservableState warned;
 
-		private Indicator(JComponent component, ObservableState valid, ObservableState warned) {
+		private Indicator(JComponent component, ObservableState invalid, ObservableState warned) {
 			this.component = requireNonNull(component);
-			this.valid = valid;
+			this.invalid = invalid;
 			this.warned = warned;
 			this.uiComponentKey = initializeUiComponentKey();
 			if (componentSupported(uiComponentKey)) {
 				component.addPropertyChangeListener("UI", event -> configureColors());
-				valid.addConsumer(state -> update());
-				warned.addConsumer(state -> update());
+				invalid.addListener(this::update);
+				warned.addListener(this::update);
 				update();
 			}
 		}
 
 		private void update() {
-			boolean enabled = component.isEnabled();
-			boolean invalid = !valid.is();
-			boolean warning = warned.is();
-			SwingUtilities.invokeLater(() -> {
-				if (invalid) {
-					component.setBackground(invalidBackgroundColor);
-				}
-				else if (warning) {
-					component.setBackground(warnedBackgroundColor);
-				}
-				else {
-					component.setBackground(enabled ? backgroundColor : inactiveBackgroundColor);
-				}
-			});
+			Color background = background();
+			SwingUtilities.invokeLater(() -> component.setBackground(background));
+		}
+
+		private @Nullable Color background() {
+			if (invalid.is()) {
+				return invalidBackgroundColor;
+			}
+			else if (warned.is()) {
+				return warnedBackgroundColor;
+			}
+			else {
+				return component.isEnabled() ? backgroundColor : inactiveBackgroundColor;
+			}
 		}
 
 		private void configureColors() {